Move external api types out of common types package
