+++ packages/common/commonTypes/src/index.ts
@@ -1,35 +1,21 @@
 import {
-  PDSResponse,
   PatientDetails,
   PatientDetailsAddress,
   PrescriptionStatusCategories,
   PrescriptionStatus,
   TreatmentType,
-  SearchParams,
   PrescriptionAPIResponse,
-  PatientAPIResponse,
   PrescriptionSummary,
-  SearchResponse,
-  IntentMap,
-  STATUS_CATEGORY_MAP,
-  PRESCRIPTION_DISPLAY_LOOKUP,
-  statusCodeMap
+  SearchResponse
 } from "./prescriptionList"
 
 export {
-  PDSResponse,
   PatientDetails,
   PatientDetailsAddress,
   PrescriptionStatusCategories,
   PrescriptionStatus,
   TreatmentType,
-  SearchParams,
   PrescriptionAPIResponse,
-  PatientAPIResponse,
   PrescriptionSummary,
-  SearchResponse,
-  IntentMap,
-  STATUS_CATEGORY_MAP,
-  PRESCRIPTION_DISPLAY_LOOKUP,
-  statusCodeMap
+  SearchResponse
 }
