+++ packages/cpt-ui/src/components/PatientDetailsBanner.tsx
@@ -27,6 +27,7 @@ export default function PatientDetailsBanner() {
     }
 
     const constructAddress = (address: PatientDetailsAddress) => {
+        const units = [address.line1, address.line2, address.city, address.postcode].filter((x) => {return x})
         const stringified = "" +
             `${address.line1 ?? ""}, ` +
             `${address.line2 ?? ""}, ` +
